Day 1 complete
Due to how importing in go works and my plugin system, I am switching to
using the embed package which will bundle the input file into the binary
at compile time. Without this approach, the runner.go script will look for the input
files relative to "runner.go" at runtime, and not in the "day*" folder, which is
not correct. Using the "embed" approach makes it so that everything with pathing is
resolved at compile time, and not runtime, and will work in any
scenario.